Create server_mock.js
This file contains a simulation of connecting to the Sensation software using Socket.IO library. The code simulates the transfer of data between the server and clients using the Socket.IO library. The server listens for incoming connections and initializes a sequence number for each client. The simulation includes randomly generated data for serial number, coordinates, and group number, which are updated every 3milliseconds and sent to each connected client.