Create client.js
This code connects to a server using socket.io-client, listens for a "seq-num" event, and updates a JSON data file with the received data. Specifically, when a "seq-num" event is received, the code checks if the message is not 1 and then updates the first point in the "points" array of the data file with the received message. The updated data is then written back to the JSON file.