Create Follow.js
This is a JavaScript code that is used for processing and displaying drone tracking data. The code uses the Exceljs and XLSX libraries to read and write data from and to Excel files respectively. It also uses the OpenLayers library to display the tracking data on a map.

The code starts by defining an array called "firstRow" that contains the column headings of the Excel file that will be read. It then adds an event listener to the "input" element in the HTML file that triggers the "onchange" function when a file is selected for reading.

The "onchange" function first checks if the "realTimeState" is true and if so, it displays an alert message telling the user to stop and save real-time data before reading drone tracking data. If the "realTimeState" is false, the function reads the selected file using the FileReader and Exceljs libraries.

Once the file is loaded, the function iterates through each sheet of the workbook and searches for the row that contains the column headings defined in the "firstRow" array. If the file is of the correct format, it sets the "typeOfFile" variable to either "readyFile" or "newFile" depending on whether the height data has already been calculated or not. If the file is not of the correct format, an alert message is displayed and the file input is reset to null.

If the file is of the correct format, the function iterates through each row of the sheet starting from the row where the column headings are found. If the "typeOfFile" is "newFile", it creates an empty "dataArr" array if the array is not already empty because real-time data has been saved. If the row is not empty, the function extracts the data from the row and sends it for processing and display on the map using the "put_point_on_map" function.

The function also includes a "drawLines" function that is used to draw lines between the points on the map after all the rows have been processed. This is used to show the path taken by the drone.

The code also includes a "DownloadFile" function that is triggered when the user clicks on the "writeToExel" button. This function checks if there is data in the map and whether the "typeOfFile" is "readyFile". If there is no data or the file has already been saved, an alert message is displayed. Otherwise, the function sorts the data in the "dataArr" array in chronological order and creates an Excel file using the XLSX library. The file is then downloaded by the user.

The code also includes a "sortFunction" that is used to sort the data in the "dataArr" array in chronological order and a "row_to_arr" function that is used to extract the data from a row in the Excel file and store it in an array